完成至p88

+++ mall-product/src/main/java/com/katzenyasax/mall/product/entity/ProductAttrValueEntity.java
@@ -43,10 +43,10 @@ public class ProductAttrValueEntity implements Serializable {
 	/**
 	 * 顺序
 	 */
-	private Integer attrSort;
+	private Long attrSort;
 	/**
 	 * 快速展示【是否展示在介绍上；0-否 1-是】
 	 */
-	private Integer quickShow;
+	private Long quickShow;
 
 }
